Added detection of unsupported sectioned datafile format.
This sectioned format is new for new, never heard about it until today.
It is a for runner of the segmented datafile format (I think) of
the approx. 40000 files on our server, we have only 4 of these.
So I made the decision not to put any effort in it until we find
that people have actually used it in measurements.  For now we
detect it, and refuse to load it.