When writing files I forgot to rename them, resulting in incompatible files.
When loading CytoSense datafiles, the DLL name is mapped to the new CyzFile
DLL, but when writing new datafiles I did not perform the reverse mapping.
The result was that newly generated files could not be read by older
software.  I added translation of the DLL name, and added some test to
make sure the translation works both ways.