Fix #1334 in CytoClus, multiple default sets when loading a workspace.
Recently the behavior of the setlist changed so it will always have a
default set on creation, but the deserialization from Xml was not
changed.  It still loaded the default set from the XML, and as a result
you got multiple default sets.  Now instead of creating a new Default set
we retrieve the existing one from the list, and let it load its properties
from the XML.