Redo of Fix Setlist serialnumber being empty #1405
Turns out the datafile property was empty in all cytosets since cytoclus 4 or atleast since we switched to workspacecontroller/Multiple files per workspace. When loading a set a new set is already made with the serialnumber of the workspace. In some cases the empty serialnumber of the set overwrote it again with empty. Now we check if its not empty and the same. Empty or different => don't overwrite. It shouldnt be able to different. Empty is possible and should be fixed here.