Added tests for serializing CalibrationPoints for old datafiles.
And changed a Debug.Assertion, there is an issue with the .Net libraries
that makes us generate datafiles that are not backwards compatible for
some (very old) instruments and the assertion triggers on that.
Unfortunately without a fix in the .net libraries there is not much we
can do about it.  dotnet/runtime#108444