Disable deprecated warning for EnableAutoCrop variable
We will enable it again (remove the disable) once enough time has passed and we can safely remove the EnableAutocrop variable from the main file. For now there are probably some files that use it and are still recent/relevant. The deprecated tag is there to make sure the iif version is used instead in future use.