[build] Update build process so we do not need Directory.build.props per folder.
All the settings are now handled in the root Directory.build.props file,
so there is no risk of forgetting one when adding a new project.
NOTE: This module can also build standalone, so we need one here
at the root level of the module.  It checks if it is the root level,
if so it setts the properties, if not it loads the higher level one.