Add number of failed crops as a datapoint to the measurementInfo
If a particle has an image and autocrop is enabled, the image is cropped during acquisition. If the contour detection on the image cannot find a particle in the image (no contour), the image is thrown away. This number represents how often that happenened during a measurement. Also fix some indenting