[CyzFile] Bugfix calculating relative measurement end time for some VERY old files.
In some very old (2010 and older) files, certain data was not present. This
meant we needed to calculate the end time in a different way.  This was
not always handled correctly.  We found this while adding unit tests before
a new change in time calculation.  This change fixes the issue.