Removed Gas Vacuole Module Support
This module no longer exists and we will probably never build one again.
There was some code for this module where I was making changes for the
new Sub Deep protocol and after discussing with George I decided to
remove it completely.

The settings are still left in the datafile for backwards compatibility
but they are marked obsolete, and cannot be used for new measurement.